\name{deleteActivity}
\alias{deleteActivity}
\title{deleteActivity function}
\usage{
deleteActivity(activityId)
}
\arguments{
  \item{activityId}{A Garmin Connect activity id}
}
\description{
Delete an activity from the local database.
}
\examples{
\dontrun{
deleteActivity(433485172)
}
}
\keyword{data}
\keyword{fitness,}

