\name{listGarminConnectActivity}
\alias{listGarminConnectActivity}
\title{listGarminConnectActivity function}
\usage{
listGarminConnectActivity(GCuser = "stats290_test",
  GCpassword = "stats290_test")
}
\arguments{
  \item{GCuser}{Garmin Connect username}

  \item{GCpassword}{Garmin Connect password}
}
\description{
List activities stored in Garmin Connect. Returns a data
frame.
}
\examples{
\dontrun{
listGarminConnectActivity()
}
}
\keyword{data}
\keyword{fitness,}

