\name{listLoadedActivity}
\alias{listLoadedActivity}
\title{listLoadedActivity function}
\usage{
listLoadedActivity()
}
\description{
List all the activities already loaded into the package
internal database. Returns a data frame.
}
\examples{
\dontrun{
listLoadedActivity()
}
}
\keyword{data}
\keyword{fitness,}

