\name{mapActivity}
\alias{mapActivity}
\title{mapActivity function}
\usage{
mapActivity(activityId)
}
\arguments{
  \item{activityId}{A Garmin Connect activity id}
}
\description{
Plot all the trackpoints of a given activity on a Google
Map. The activity has to be available in the local
database.
}
\examples{
\dontrun{
mapActivity(439915418)
}
}
\keyword{data}
\keyword{fitness,}

